+++ pipen/exceptions.py
@@ -62,5 +62,5 @@ class ConfigurationError(PipenException):
     """When something wrong set as configuration"""
 
 
-class ProcWorkdirConflictException(PipenException):
+class PipenOrProcNameError(PipenException):
     """ "When more than one processes are sharing the same workdir"""