+++ tests/test_pipen.py
@@ -5,6 +5,7 @@ from pipen.exceptions import (
     ProcDependencyError,
     PipenSetDataError,
 )
+from pipen.proc import PipenOrProcNameError
 
 from .helpers import (
     ErrorProc,
@@ -156,3 +157,26 @@ def test_subclass_pipen(tmp_path, caplog):
         ...
 
     assert MyPipe2().name == "MyPipe2"
+
+
+def test_invalid_name():
+    class MyPipe3(Pipen):
+        name = "a+"
+
+    with pytest.raises(PipenOrProcNameError, match="Invalid pipeline name"):
+        MyPipe3().run()
+
+
+def test_duplicate_proc_name():
+    class MyProc1(Proc):
+        ...
+
+    class MyProc2(Proc):
+        requires = MyProc1
+        name = "MyProc1"
+
+    class MyPipe4(Pipen):
+        starts = MyProc1
+
+    with pytest.raises(PipenOrProcNameError, match="already used by another"):
+        MyPipe4().run()